Use initial_conditions to initialize cpp data

+++ micro-apps/p3/types.hpp
@@ -2,6 +2,6 @@
 #define INCLUDE_TYPES
 
 typedef int Int;
-typedef float Real; // for now; see issue #11
+typedef double Real; // for now; see issue #11
 
 #endif
